scripts/CustomerSpecific/PFF/Snapv3/requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
moviepy==1.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
sort-track==0.0.7
clarifai==11.7.2
scipy==1.11.2
python-dotenv==1.0.0